requirements.txt
streamlit>=1.30.0
pandas>=2.1.1
numpy>=1.26.0
altair>=5.1.2
plotly==5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
pillow>=10.0.1
requests>=2.31.0
typing-extensions>=4.9.0
scipy>=1.11.3
snowflake-connector-python>=3.5.0
snowflake-snowpark-python>=1.10.0